Fix popups and semi-support older builds
Now requires storage access, because we need to pull down the data files from ffbeEquip and use those to populate the data from older links.  This is done on a daily basis, on demand.

Needed to pull in `browser-polyfill` because all browsers aren’t supporting the same APIs, which is annoying but, whatever.

Move to non-minified version of the `exportBuildAsImage` script, because it is easier to debug.

Also make the Firefox extension as part of the general build process.